+++ src/vue/pages/VerificationGeneral.vue
@@ -15,6 +15,10 @@
 import GeneralKycForm from '@/vue/forms/GeneralKycForm'
 import { Bus } from '@/js/helpers/event-bus'
 
+const EVENTS = {
+  kycRecoverySubmit: 'kyc-recovery-submit',
+}
+
 export default {
   name: 'verification-general-form',
   components: {
@@ -24,6 +28,7 @@ export default {
   methods: {
     showSuccessMessage () {
       Bus.success('verification-form.request-submitted-msg')
+      this.$emit(EVENTS.kycRecoverySubmit)
     },
   },
 }